requirements.txt
# Core Streamlit and Web Framework
streamlit>=1.28.0
python-dotenv>=1.0.0

# AI and Machine Learning
transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.99

# Document Processing
PyPDF2>=3.0.1
pdfplumber>=0.9.0
python-docx>=0.8.11

# PDF Generation and Reporting
reportlab>=4.0.4

# Security and Utilities
cryptography>=41.0.0
psutil>=5.9.0

# Additional utilities for text processing
nltk>=3.8
regex>=2023.6.3